University Management(Switch)
Q2: Create a C program for a university registration system. The program will first ask the user to enter a registration code. The registration code should be a variable whose value is your student ID number. Check if the user-entered code matches with the pre-defined registration code or not. If the code matches, the system will show the following options:

1. Register for courses
 2. View student details Enter any other number to exit.
Registering for courses: Allow the user to choose from available courses and register them. Each course should have a name. Display a message confirming the registration.
                                                                                            
Viewing student details: Display the details of the student including the registered course,
The student may register for only one course, and while viewing details of the course, the last
registered course will be displayed.

The program should loop until the user chooses to exit.
Ensure to handle cases where the user enters an invalid option or if no course is registered.

The program should not use arrays or functions.