Bank_Management.c
Q2: Create a C language program for a banking system. The program will first ask the user to enter a PIN. The PIN should be a variable whose value is your roll number. Check if the user entered PIN matches with the pre-defined PIN or not. If the PIN matches, the system will show the following options:
 1. Check balance. 2. Transfer to another bank. 3. Change PIN.

Transferring to another bank will ask for a Bank account number and then amount. If the amount exceeds the balance, prompt the user to enter the option again with the message, "Amount exceeding the current balance." Ask the user to re-enter the PIN and verify the PIN again before deducting the amount from the balance. Show the new balance to the user after deduction.

In order to change the PIN, the user will re-enter the existing PIN and verify. If the PIN matches then ask the user to enter the new PIN and update the value in the pre-defined variable. If the PIN does not match during verification, exit the program.

The program will keep running in a loop. If anywhere in the code, the user entered PIN does not match with the Pre-defined PIN, exit the program.