Create dna_neuclotide_count.c
Scenario: Deoxyribonucleic acid (DNA) is an information carrier in humans’ cells that helps in the 
development and functioning of an organism.  The information in DNA is stored as a code made up of four 
chemical bases (letters) or nucleotides: adenine (A), guanine (G), cytosine (C), and thymine (T). Human 
DNA consists of about 3 billion bases, and more than 99 percent of those bases are the same in all people. 
The meaning of the DNA code lies in the sequence of these nucleotides’ letters i.e. A, G, C & T.
 Task:
 Read a long DNA sequence up to 20 nucleotides and display a histogram of each letter by printing number 
of ‘*’ against each letter.   
Sample Input: 
AGCTTACATATTACGAGTTG
Sample Output: 
A: * * * * * *
G: * * * * 
C: * * *
T:  * * * * * * *